requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sqlalchemy==2.0.23
